MEDIUM: acme: use lowercase for challenge names in configuration
Both the RFC and the IANA registry refers to challenge names in
lowercase. If we need to implement more challenges, it's better to
use the correct naming.

In order to keep the compatibility with the previous configurations, the
parsing does a strcasecmp() instead of a strcmp().

Also rename every occurence in the code and doc in lowercase.

This was discussed in issue #1864

(cherry picked from commit 84589a9f48963efe76a4c6dcd21b8dbc71c296e0)
Signed-off-by: William Lallemand <wlallemand@haproxy.com>