BUG/MEDIUM: mux-spop: Reject connection attempts from a non-spop frontend
It is possible to crash the process by initializing a connection to a SPOP
server from a non-spop frontend. It is of course unexpected and invalid. And
there are some checks to prevent that when the configuration is
loaded. However, it is not possible to handle all cases, especially the
"use_backend" rules relying on log-format strings.

It could be good to improve the backend selection by checking the mode
compatibility (for now, it is only performed for the HTTP).

But at the end, this can also be handled by the SPOP multiplexer when it is
initialized. If the opposite SD is not attached to an SPOE agent, we should
fail the mux initialization and return an internal error.

This patch must be backported as far as 3.1.

(cherry picked from commit 886a248be4d480a095b6aac40b28d8386b79d9c2)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 35286e7584ec56fd5ec09c2092634427d6310a85)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>