BUG/MEDIUM: spoe: Improve error detection in SPOE applet on client abort
It is possible to interrupt a SPOE applet without reporting an error. For
instance, when the client of the parent stream aborts. Thanks to this patch,
we take care to report an error on the SPOE applet to be sure to interrupt
the processing. It is especially important if the connection to the agent is
queued. Thanks to 886a248be ("BUG/MEDIUM: mux-spop: Reject connection
attempts from a non-spop frontend"), it is no longer an issue. But there is
no reason to continue to process if the parent stream is gone.

In addition, in the SPOE filter, if the processing is interrupted when the
filter is destroyed, no specific status code was set. It is not a big deal
because it cannot be logged at this stage. But it can be used to notify the SPOE
applet. So better to set it.

This patch should be backported as far as 3.1.

(cherry picked from commit 49db9739d0591d15907e0241f3bcc8e7adb4fd3a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ec271a05a175ddf1c22841c45658e5a7f299d032)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>