MINOR: debug: report the process id in warnings and panics
Warning and panic messages currently do not report the PID. This is
annoying when trying to reproduce problems because warnings do not
allow know which process to attach to in order to debug, and panics
do not permit to know which core dump corresponds to which dump.
Let's add them in both messages. This should probably be backported
at least to 3.2.

(cherry picked from commit 8153cf1e51fb1b722ab2b16a452c043fbf307b56)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>