DEBUG: stream: count the number of passes in the connect loop
Normally the connect loop cannot loop, but some recent traces can easily
convince one of the opposite. Let's add a counter, including in panic
dumps, in order to avoid the repeated long head scratching sessions
starting with "and what if...". In addition, if it's found to loop, this
time it will be certain and will indicate what to zoom in. This should
be backported to 3.2.

(cherry picked from commit c3f94fbd9b49a140c10dd6036585d3b76779feb4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>